Create wrapper GET/POST methods (major change)
This change is to remove the code duplication of the decoding of
get/post request from the ua.

This way, roles only call one method (_fetch) implementing all
that's required in order to return the call's result
(either decoded or just the status).

Also, it is not longer required to pass the token to the calls,
we KNOW we need it so it's added by default unless specifically
requested not to (no_token => 1).