Return a User object from the login methods.
Better connect the API / account through a User object, which
will be used for most of the API interaction (few methods do
not involve login and can be called directly from the API module)

This requires the passing of token to the API object where
GET/POST are implemented.

Also, simplified the params/return to/from the GET/POST methods.